configure: rename modules
