+++ quoteengine/quote.py
